chore: Remove stale reporter code, switch package to use new IO protocol
duckdb and the base class were unused. Since we need to employ different
abstractions for different IO models, a general base class was not that
useful aside for maybe type safety.
