+++ src/nnbench/__init__.py
@@ -1,7 +1,7 @@
 """A framework for organizing and running benchmark workloads on machine learning models."""
 
 from .core import benchmark, parametrize, product
-from .reporter import BenchmarkReporter, ConsoleReporter, FileReporter
+from .reporter import ConsoleReporter, FileReporter
 from .runner import collect, run
 from .types import Benchmark, BenchmarkFamily, BenchmarkRecord, Parameters
